+++ ddtrace/contrib/aioredis/patch.py
@@ -5,6 +5,11 @@ import sys
 import aioredis
 
 from ddtrace import config
+from ddtrace._trace.utils_redis import ROW_RETURNING_COMMANDS
+from ddtrace._trace.utils_redis import _run_redis_command_async
+from ddtrace._trace.utils_redis import _trace_redis_cmd
+from ddtrace._trace.utils_redis import _trace_redis_execute_pipeline
+from ddtrace._trace.utils_redis import determine_row_count
 from ddtrace.internal.constants import COMPONENT
 from ddtrace.internal.utils.wrappers import unwrap as _u
 from ddtrace.pin import Pin
@@ -25,11 +30,6 @@ from ...internal.utils.formats import CMD_MAX_LEN
 from ...internal.utils.formats import asbool
 from ...internal.utils.formats import stringify_cache_args
 from .. import trace_utils
-from ..trace_utils_redis import ROW_RETURNING_COMMANDS
-from ..trace_utils_redis import _run_redis_command_async
-from ..trace_utils_redis import _trace_redis_cmd
-from ..trace_utils_redis import _trace_redis_execute_pipeline
-from ..trace_utils_redis import determine_row_count
 
 
 try:
