+++ ddtrace/contrib/flask_cache/utils.py
@@ -1,7 +1,8 @@
 # project
+from ddtrace._trace.utils_redis import _extract_conn_tags as extract_redis_tags
+
 from ...ext import net
 from ..pylibmc.addrs import parse_addresses
-from ..trace_utils_redis import _extract_conn_tags as extract_redis_tags
 
 
 def _resource_from_cache_prefix(resource, cache):
